pyproject.toml
[project]
name = "proj1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "accelerate>=1.12.0",
    "customtkinter>=5.2.2",
    "diffusers>=0.36.0",
    "numpy>=2.4.2",
    "pillow>=12.1.1",
    "torch>=2.10.0",
    "transformers>=5.2.0",
]
